change it to add the logged in user when creating a donation
